5.1.10 Перенесет рендеринг доски в controllers/page.js
