+++ src/main.js
@@ -1,116 +1,13 @@
+import PageController from "./controllers/page.js";
 import ProfileComponent from "./components/profile.js";
 import MainNavigationComponent from "./components/main-navigation.js";
 import SortingComponent from "./components/sorting";
-import BoardComponent from "./components/board.js";
-import FilmsListComponent from "./components/films-list.js";
-import ExtraInfoAboutFilmsComponent from "./components/extra-info-about-films.js";
-import FilmComponent from "./components/film.js";
-import FilmDetailsComponent from "./components/film-details.js";
-import ShowMoreButtonComponent from "./components/show-more-button.js";
 import FilmsAmountComponent from "./components/films-amount.js";
-import NoFilmsComponent from "./components/no-films.js";
 import {generateFilms, generateWatchedFilmsCount, generateFilmsAmount} from "./mock/film.js";
 import {generateFilters} from "./mock/filter.js";
-import {render, append, remove} from "./utils/render.js";
+import {render} from "./utils/render.js";
 
 const FILM_COUNT = 15;
-const SHOWING_FILMS_COUNT_ON_START = 5;
-const SHOWING_FILMS_COUNT_BY_BUTTON = 5;
-const SHOWING_TOP_RATED_FILMS_COUNT = 2;
-const SHOWING_MOST_COMMENTED_FILMS_COUNT = 2;
-
-const renderFilm = (containerElement, film) => {
-  const openFilmDetailsPopup = () => {
-    bodyElement.classList.add(`hide-overflow`);
-    append(bodyElement, filmDetailsComponent);
-    document.addEventListener(`keydown`, onEscKeyDown);
-    filmDetailsComponent.setFilmDetailsCloseButtonClickHandler(closeFilmDetailsPopup);
-  };
-
-  const closeFilmDetailsPopup = () => {
-    bodyElement.classList.remove(`hide-overflow`);
-    remove(filmDetailsComponent);
-    document.removeEventListener(`keydown`, onEscKeyDown);
-  };
-
-  const onEscKeyDown = (evt) => {
-    const isEscKey = evt.key === `Escape` || evt.key === `Esc`;
-
-    if (isEscKey) {
-      closeFilmDetailsPopup();
-    }
-  };
-
-  const bodyElement = document.querySelector(`body`);
-
-  const filmComponent = new FilmComponent(film);
-
-  filmComponent.setFilmTitleClickHandler(openFilmDetailsPopup);
-  filmComponent.setFilmPosterClickHandler(openFilmDetailsPopup);
-  filmComponent.setFilmCommentsClickHandler(openFilmDetailsPopup);
-
-  const filmDetailsComponent = new FilmDetailsComponent(film);
-
-  filmDetailsComponent.setFilmDetailsCloseButtonClickHandler(closeFilmDetailsPopup);
-
-  render(containerElement, filmComponent);
-};
-
-const renderFilmsList = (containerElement, films, showingFilmsCount) => {
-  const filmsListComponent = new FilmsListComponent();
-  const filmsListElement = filmsListComponent.getElement();
-  const filmsListContainerElement = filmsListElement.querySelector(`.films-list__container`);
-
-  films.slice(0, showingFilmsCount)
-    .forEach((film) => renderFilm(filmsListContainerElement, film));
-
-  const showMoreButtonComponent = new ShowMoreButtonComponent();
-
-  render(filmsListElement, showMoreButtonComponent);
-
-  showMoreButtonComponent.setClickHandler(() => {
-    const prevFilmsCount = showingFilmsCount;
-    showingFilmsCount = showingFilmsCount + SHOWING_FILMS_COUNT_BY_BUTTON;
-
-    films.slice(prevFilmsCount, showingFilmsCount)
-      .forEach((film) => renderFilm(filmsListContainerElement, film));
-
-    if (showingFilmsCount >= films.length) {
-      remove(showMoreButtonComponent);
-    }
-  });
-
-  render(containerElement, filmsListComponent);
-};
-
-const renderExtraFilmsList = (containerElement, title, films, showingFilmsCount) => {
-  const extraInfoAboutFilmsComponent = new ExtraInfoAboutFilmsComponent(title);
-  const extraInfoAboutFilmsElement = extraInfoAboutFilmsComponent.getElement();
-  const filmsListContainerElement = extraInfoAboutFilmsElement.querySelector(`.films-list__container`);
-
-  films.slice(0, showingFilmsCount)
-    .forEach((film) => renderFilm(filmsListContainerElement, film));
-
-  render(containerElement, extraInfoAboutFilmsComponent);
-};
-
-const renderBoard = (containerElement, films) => {
-  if (films.length === 0) {
-    render(containerElement, new NoFilmsComponent());
-    return;
-  }
-
-  const boardComponent = new BoardComponent();
-  const boardElement = boardComponent.getElement();
-  const filmsSortedByRating = films.slice().sort((a, b) => b.rating - a.rating);
-  const filmsSortedByComments = films.slice().sort((a, b) => b.comments.length - a.comments.length);
-
-  renderFilmsList(boardElement, films, SHOWING_FILMS_COUNT_ON_START);
-  renderExtraFilmsList(boardElement, `Top rated`, filmsSortedByRating, SHOWING_TOP_RATED_FILMS_COUNT);
-  renderExtraFilmsList(boardElement, `Most commented`, filmsSortedByComments, SHOWING_MOST_COMMENTED_FILMS_COUNT);
-
-  render(containerElement, boardComponent);
-};
 
 const siteHeaderElement = document.querySelector(`.header`);
 const siteMainElement = document.querySelector(`.main`);
@@ -124,5 +21,8 @@ const filmsAmount = generateFilmsAmount();
 render(siteHeaderElement, new ProfileComponent(watchedFilmsCount));
 render(siteMainElement, new MainNavigationComponent(filters));
 render(siteMainElement, new SortingComponent());
-renderBoard(siteMainElement, films);
+
+const pageController = new PageController(siteMainElement);
+pageController.render(films);
+
 render(footerStatisticsElement, new FilmsAmountComponent(filmsAmount));
